Make Init() run migrations with retries
Since the docker compose file we use starts with calling the "init"
command on every startup to handle DB migrations, lets try reaching the
database with retries just in case the DB is temporarily unavailable
(e.g. both machines have been rebooted at the same time and are in the
process of starting up).

Beacuse the migrations are handled in the separate migrations package create a new
"managerutils" package for holding the RetryWithBackoff() code so we can
use it in both the server and migration packages. Also make it possible to
pass down the initCtx to migrations.Up() so we can get a timeout which
is now also passed on to goose. Finally move tests for the retry code to
the managerutils package. Include fix for panic when sleepBase is
set to 1ns including a new test that triggers it as well as some
additional tweaks.

Also increase initCtx timeout to give the retries time to work.
